more db tests
