+++ rocket-backend/internal/database/settings_table.go
@@ -63,34 +63,3 @@ func (s *service) GetSettingsByUserID(userID uuid.UUID) (*types.Settings, error)
 
 	return &settings, nil
 }
-
-func (s *service) GetUserImage(userID uuid.UUID) (*types.UserImage, error) {
-	query := `
-		SELECT
-			i.id, i.image_name, i.image_data
-		FROM
-			settings s
-		JOIN
-			image_store i ON s.image_id = i.id
-		WHERE
-			s.user_id = $1
-	`
-
-	var img types.UserImage
-	err := s.db.QueryRow(query, userID).Scan(
-		&img.ID,
-		&img.Name,
-		&img.Data,
-	)
-
-	if err != nil {
-		if err == sql.ErrNoRows {
-			logger.Warn("No image found for user:", userID)
-			return nil, nil
-		}
-		logger.Error("Failed to get user image", err)
-		return nil, fmt.Errorf("failed to get user image: %w", err)
-	}
-
-	return &img, nil
-}
